+++ sdk/resourcemanager/baremetalinfrastructure/armbaremetalinfrastructure/CHANGELOG.md
@@ -1,5 +1,31 @@
 # Release History
 
+## 2.0.0 (2023-05-12)
+### Breaking Changes
+
+- Type of `ErrorResponse.Error` has been changed from `*ErrorDefinition` to `*ErrorDetail`
+- Struct `ErrorDefinition` has been removed
+
+### Features Added
+
+- New enum type `ProvisioningState` with values `ProvisioningStateAccepted`, `ProvisioningStateCanceled`, `ProvisioningStateCreating`, `ProvisioningStateDeleting`, `ProvisioningStateFailed`, `ProvisioningStateMigrating`, `ProvisioningStateSucceeded`, `ProvisioningStateUpdating`
+- New function `NewAzureBareMetalStorageInstancesClient(string, azcore.TokenCredential, *arm.ClientOptions) (*AzureBareMetalStorageInstancesClient, error)`
+- New function `*AzureBareMetalStorageInstancesClient.Create(context.Context, string, string, AzureBareMetalStorageInstance, *AzureBareMetalStorageInstancesClientCreateOptions) (AzureBareMetalStorageInstancesClientCreateResponse, error)`
+- New function `*AzureBareMetalStorageInstancesClient.Delete(context.Context, string, string, *AzureBareMetalStorageInstancesClientDeleteOptions) (AzureBareMetalStorageInstancesClientDeleteResponse, error)`
+- New function `*AzureBareMetalStorageInstancesClient.Get(context.Context, string, string, *AzureBareMetalStorageInstancesClientGetOptions) (AzureBareMetalStorageInstancesClientGetResponse, error)`
+- New function `*AzureBareMetalStorageInstancesClient.NewListByResourceGroupPager(string, *AzureBareMetalStorageInstancesClientListByResourceGroupOptions) *runtime.Pager[AzureBareMetalStorageInstancesClientListByResourceGroupResponse]`
+- New function `*AzureBareMetalStorageInstancesClient.NewListBySubscriptionPager(*AzureBareMetalStorageInstancesClientListBySubscriptionOptions) *runtime.Pager[AzureBareMetalStorageInstancesClientListBySubscriptionResponse]`
+- New function `*AzureBareMetalStorageInstancesClient.Update(context.Context, string, string, Tags, *AzureBareMetalStorageInstancesClientUpdateOptions) (AzureBareMetalStorageInstancesClientUpdateResponse, error)`
+- New function `*ClientFactory.NewAzureBareMetalStorageInstancesClient() *AzureBareMetalStorageInstancesClient`
+- New struct `AzureBareMetalStorageInstance`
+- New struct `AzureBareMetalStorageInstanceProperties`
+- New struct `AzureBareMetalStorageInstancesListResult`
+- New struct `ErrorAdditionalInfo`
+- New struct `ErrorDetail`
+- New struct `StorageBillingProperties`
+- New struct `StorageProperties`
+
+
 ## 1.1.0 (2023-03-27)
 ### Features Added